Add files via upload
The part of the realized codes by using Dynamo®, the CAD drawings (standard floor and roof) of two structures in the case study, the element prototypes and the generated Revit® BIM model (standard floor and roof) by using this methodology are available. It need Dynamo 2.0 or higher version, and the Dynamo pakages "Clockwork for Dynamo 2.0" and "Data shapes" are needed to successfully execute these node flows.